BEISDEAS-232 Change scrolling approach to just scroll on click, instead of using angular routing.
Using angular hash fragments to scroll to an id on the page is poorly supported at the moment, so you have to manually find the element using the fragment and then scroll to it. The event which fires when a fragment changes triggers many more times than is desired, including when other things are clicked on the page. For this reason, we should instead use a simple click event to scroll the page, without using the angular router