BEISDEAS-259 Fix back angular router on the encyclopedia pages.
Some of the current encyclopedia page links had '&'s in them. Of course this is a special character in urls e.g. /your-home/windows&doors

The Angular router only saves the bit up to the '&' in the browser's history. When you press back, you go to /your-home/windows. The your-home component looks for a encyclopedia page called "windows", fails to find it, and then returns you to the homepage. This redirect also causes the forward button to grey out

I've applied a straight-forward fix to rename the routes to e.g. /your-home/windows-and-doors