BEISDEAS-258 Change error handling for postcodes.
If there is an error with the postcode lookup, the postcode-lookup component will now display an error message and set the postcode to null.
The EPC lookup checks for this null value, and will display the continue button. This means that users are notified of the error, but can continue on their journey