BEISDEAS-304: Add analytics for Facebook ads
Extract the referrer ID from a query parameter when the app first loads,
and send it to Google Analytics.

Referrer ID query strings will be added to Facebook ads (and potentially
other sources) so that we can correlate user behaviour with how they
found the site in the first place. For example, it will let us see
whether changing the wording of a Facebook ad leads increases the
proportion of users who make it to the plan page.

The dimension reported to Google Analytics is a session dimension, which
means that it is linked to all of a user's actions on the site within
that session.