BEISDEAS-304: Fix syntax for unsubscribing from query parameter events
When we inspect the query parameters to see if the user has been
referred from a Facebook ad, we only want to check the parameters on the
first page load.

The previous attempt at this attempted to unsubscribe after processing
the first set of query parameters, but this didn't work due to a
misunderstanding of Rxjs observables.

In this commit, we ignore the first query parameter change - it looks
like there's always a queryParamMap change event when the app loads, and
the map is always blank. The second event contains the map we're
interested in, and we check it for the referrerid.